requirements.txt
nose==1.3.7
python-magic==0.4.10
wheel==0.24.0
